setup dynamic displays for admins
