+++ src/components/admin-components/ShowOnDemand.js
@@ -1,0 +1,8 @@
+import React from "react";
+
+function ShowOnDemand({ show, children }) {
+  if (!show) return <></>;
+  return <>{children}</>;
+}
+
+export default ShowOnDemand;
